LuaJ LuaThread and Globals change
Discovered the issue behind this bug when experimenting with debug hooks and found that not only is it that 1 hook applies to a thread, but 1 hook applies to all scripts. When looking into the source I found that LuaJ's LuaThread kept debug info in the thread's state. And because of how Each (real) thread shares globals they also all shared this state.

Debug info will now be accurate regardless of how many (real) threads you have running.

Untested: coroutines in theory should work as expected still.

Possible things broken:
Anywhere that a script is run in a *new thread* but doesn't set the AdvancedMacros.globals. luaThread info